requirements.txt
numpy
aiortc
cryptography
dataclasses # backport to python 3.6
dpcontracts
flask
forbiddenfruit>=0.1.3
loguru
matplotlib
nest_asyncio
packaging
pandas
protobuf
pydagogy
PyNaCl
requests
scikit-learn
seaborn
sqlitedict
statsmodels
typeguard
typing-extensions # backport to older python 3
websockets
jupyter
pysyft
torch
TenSEAL

